Add reusable task to configure DNS resolver
Add a reusable task in order to set unbound as a local resolver
and adjust the `/etc/resolv.conf` to be a "static file" (it can be
a symlink and dynamicall generated) and to only honor local unbound.

(Strictly speaking we can just remove `/etc/resolv.conf` and by
not having it 127.0.0.1 will be the only nameserver to be queried...
this can be a possible future simplification! (I have no idea if
that semantic is also valid for possible $RANDOM containers, but I would not
expect any differences!))